change cargo metadata attribute
