+++ dev-tools/xtask/src/cmd/power_set/mod.rs
@@ -2,8 +2,10 @@ use std::collections::{BTreeMap, BTreeSet};
 
 use anyhow::Context;
 
-use crate::cmd::IGNORED_PACKAGES;
-use crate::utils::{cargo_cmd, comma_delimited, parse_features, test_package_features, XTaskMetadata};
+mod utils;
+
+use crate::{cmd::IGNORED_PACKAGES, utils::{cargo_cmd, comma_delimited}};
+use utils::{parse_features, test_package_features, XTaskMetadata};
 
 #[derive(Debug, Clone, clap::Parser)]
 pub struct PowerSet {
